remove scroll to formik, add snackbar component
